chore: extract blackbox Record to shared library
