+++ fsw/blackbox/Cargo.toml
@@ -1,8 +1,14 @@
 [package]
 name = "blackbox"
-license = "MIT OR Apache-2.0"
-edition = "2024"
-version = "0.1.0"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+publish = false
+
+[features]
+roci = ["dep:roci", "std"]
+std = ["roci/std"]
 
 [dependencies]
 zerocopy = { version = "0.8.13", features = ["derive"] }
+roci = { path = "../../libs/roci", default-features = false, optional = true }
