+++ fsw/sensor-fw/Cargo.toml
@@ -41,8 +41,8 @@ dshot-frame = "0.1.3"
 bitfield-struct = "0.9"
 modular-bitfield = "0.11"
 zerocopy = { version = "0.8", default-features = false, features = ["derive"] }
-cobs.version = "0.2.3"
-cobs.default-features = false
+cobs = { version = "0.2.3", default-features = false }
+blackbox = { path = "../blackbox" }
 
 usbd-serial = "0.1"
 usb-device = "0.2.9"
